+++ src/main/java/com/aipark/web/controller/AuthController.java
@@ -36,9 +36,4 @@ public class AuthController {
         authService.logout(tokenRequestDto);
         return "로그아웃되셨습니다.";
     }
-    @PostMapping("/change-password")
-    public void changePassword(@RequestBody MemberDto.ChangeRequest changeRequestDto){
-        authService.changePassword(changeRequestDto);
-
-    }
 }
