+++ src/main/java/com/aipark/web/controller/MemberController.java
@@ -4,9 +4,7 @@ import com.aipark.biz.service.MemberService;
 import com.aipark.web.dto.MemberDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RequiredArgsConstructor
 @RestController
@@ -17,6 +15,11 @@ public class MemberController {
     @GetMapping
     public ResponseEntity<MemberDto.MemberResponse> getMyMemberInfo(){
         return ResponseEntity.ok(memberService.getMyInfo());
+    }
+    @PostMapping("/change-password")
+    public void changePassword(@RequestBody MemberDto.ChangeRequest changeRequestDto){
+        memberService.changePassword(changeRequestDto);
+
     }
 
 }
